+++ src/events/ready.js
@@ -1,10 +1,11 @@
+const logger = require('../lib/logger');
 const { restoreVoiceTracking } = require('./voiceStateUpdate');
 
 module.exports = {
   name: 'clientReady',
   once: true,
   execute(client) {
-    console.log(`Login sebagai ${client.user.tag} di ${client.guilds.cache.size} server.`);
+    logger.info(`Login sebagai ${client.user.tag} di ${client.guilds.cache.size} server.`);
     restoreVoiceTracking(client);
   },
 };